Rename `pyramid_creation.py` into `pyramids.py`
