+++ fractal_tasks_core/tasks/illumination_correction.py
@@ -30,7 +30,7 @@ from skimage.io import imread
 from fractal_tasks_core.ome_zarr.channels import get_omero_channel_list
 from fractal_tasks_core.ome_zarr.channels import OmeroChannel
 from fractal_tasks_core.ome_zarr.ngff import load_NgffImageMeta
-from fractal_tasks_core.ome_zarr.pyramid_creation import build_pyramid
+from fractal_tasks_core.ome_zarr.pyramids import build_pyramid
 from fractal_tasks_core.roi import check_valid_ROI_indices
 from fractal_tasks_core.roi import (
     convert_ROI_table_to_indices,
